Migration to the rake-pipeline
Using the rake pipeline to generate the assets for the website,
including the HTML templates. We have LESS generation, and hex digesting
for the file names of images and CSS.

Because the cache-buster filter in the rake pipeline alters the names of
the resources, we need a way to determine the new names so that they can
be included in the Slim templates.

The workaround for this is to rub a Ruby statement to look-up the newest
asset file and use that in the template. Not the prettiest, but does get
at the correct file name.

The gotcha is that the assets must be generated in the right order so
that dependencies will exist when they are looked up by the file name.
So if you had an HTML file including a CSS file, the CSS must be
generated first in the Assetfile